Added the ability to delete links from talks made some of the other tag and link related views more inline with how they should be for AJAX requests.

+++ core/views/profile_link_delete.py
@@ -1,7 +1,8 @@
+import json
 from django.shortcuts import redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ObjectDoesNotExist
-from django.http import HttpResponseNotFound
+from django.http import HttpResponse, HttpResponseNotFound, HttpResponseForbidden
 
 from core.models import UserLink
 
@@ -9,8 +10,12 @@ from core.models import UserLink
 def profile_link_delete(request, link_id):
     if request.method == "POST":
         link = get_object_or_404(UserLink, pk=link_id)
+
+        if link.user != request.user:
+            return HttpResponseForbidden()
+
         link.delete()
 
-        return redirect(request.user.get_absolute_url())
+        return HttpResponse(json.dumps({}), content_type="application/json")
     else:
         return HttpResponseNotFound()
